WIP: setup for cpm_priv_map lib outside of tool
This moves the implementation of the Rust Type'd version of the CPM
lifter and projector. The main issue is in testing it. I have started to
define the data structures but there is some comlexity in that some of
hte fields are embeded into yaml strings, but I would like to make sure
that this format is validated in parsing and project deterministically.
This led me to realize that this part of the tool truly is a *tool*
wrapper instead of the hypertool part, ie a tool that is combining and
creating a pipeline of flows between 3rd party/independent tools. All
three depend on the hypercore. This likely is one step towards the
eventual organization but should allow fast testing and rapid
prototyping.

Signed-off-by: Nathan Dautenhahn <daut@serenitix.io>