WIP:CHKPT: custom optional field types with deserialization
Each of these custom fields has the ability to be either a specific
object, list of names/objects, or the string "all". As such I wanted to
encode the structure of the grammar in the types of the data structures
so that the grammar is visible in teh rust types. This meant the use of
an enum for each type of optional field, a default value function, and a
custom deserialization routine that parses and validates the all string
or matches the struct def type and deserializes. There is plenty more
and I haven't tested yet.

Signed-off-by: Nathan Dautenhahn <daut@serenitix.io>