Add generic ObjectDomain::new interface back
This add back the generic object domain creation function that takes a
string name and a vector of objectIDs. The main goal is to keep it
general if the client wants that. It could add a check, marked as as
TODO in the code, to ensure no duplicate domains are added. This change
also fixes the test suite tests that evaluate it.

Signed-off-by: Nathan Dautenhahn <daut@serenitix.io>